refactor: make cards list a shared component
